Make the idp code less che/codewind specific plus more
