+++ idp/idp.yaml
@@ -1,0 +1,141 @@
+apiVersion: codewind.dev/v1alpha1
+kind: IterativeDevelopmentPack # I am less convinced about including this then John is
+metadata: 
+  name: java-liberty-pack # Presume this is only used as user visible text
+  language: java  # I presume this is only used as user visible text
+  framework: liberty  # I presume this is a unique ID (non user visible) for identification of the IDP template
+  version: 0.0.1 # Stolen from Appsody; semanatic version for machine use, or user-display-in-UI only?
+  maintainers:  # Stolen from Appsody
+  - name: Jonathan West
+    email: jgwest@email.com
+    github-id: jgwest
+
+spec:
+
+  dev: # TODO: Better name? local?
+    watched:
+      # On the local machine, which path to watch, from the root folder of the project.
+      # For example /codewind-workspace/my-project would be /
+      watchDir: /src # Optional, otherwise assumed to be the root /
+      ignoredPaths: # Same format as filewatcher, Optional, if not specified then no excludes.
+      - /target
+      - /some.file
+
+    uploadFilter: # Unclear if we can just combine this with watch, or if it needs to be separate
+      ignoredPaths: # Same format as filewatcher, Optional, if not specified then no excludes.
+      - /target
+      
+    typeDetection: # Optional: Rules that are used to detect if this IDP is applicable to a given project (eg OpenLiberty IDP for an OpenLiberty-based app)
+    # At least one rule must match for a type; ie this is an OR clause list, not an AND clause list.
+    - pathSelector: # At least one 'pathSelector' entry must exist if typeDetection is specified
+        # Optional, Either a 'path' xor a 'filename' must be specified 
+        rootPath: # example: /pom.xml, or /go.mod
+        filenameWildcard: # example: server.xml, or *.go; this means recursively walk a project and look for filenames that match this pattern. Same pattern style as filewatcher.
+      textStringsToMatch: # Optional: If a file matches one of the selectors, then look for at least one of these strings (ie OR clause list, not AND clause list). 
+      - net.wasdev.wlp.maven.plugins # Example: if this ID is found in the file, then this IDP should be considered to be applicable to the project
+    
+  runtime:
+    
+    image: docker.io/ibmcom/websphere-liberty:latest
+
+    endpoints: # Not optional if HTTP(S) port is specified
+      containerHealth: /health # How to tell the container is healthy
+      appRoot: /app # Not a health check
+
+    ports: # All are optional, display a warning if neither HTTP port is specified
+      internalHttpPort: 9080 
+      internalHttpsPort: 9443
+      internalDebugPort: 7777
+      internalPerformancePort: 9050
+      
+    logs: # Optional: Ability to reference arbitrary log file types that aren't included in container stderr/stdout
+    - type: some.type
+      path: /logs/(etc)
+
+    env: # Optional: Ability to map key/value pairs into the container as environment variables, these are runtime only
+    - key: value 
+    - key2: value2 
+
+    kubernetes: # Values only used for Kube deployments
+    
+      # TODO: Are there other Kube resource parameters we need to include here? securityContext? (cluster) role bindings?
+      livenessProbe: # Optional, otherwise sane defaults should be used
+        initialDelaySeconds: 15
+        timeoutSeconds: 60
+
+      readinessProbe: # Optional, otherwise sane defaults should be used
+        initialDelaySeconds: 15
+        timeoutSeconds: 60
+
+      memoryLimit: 600Mi # Arguable whether this should only be in kubernetes 
+
+    # runAsUser: 185 
+
+  shared:
+    volumes: 
+    - name: idp-data-volume # or perhaps this would always be autogenerated to prevent collisions
+      labels: idp # kube only?
+      accessModes: # kube only
+      - ReadWriteMany
+      size: 1Gi # kube only
+      
+    - name: maven-cache-volume 
+      labels: idp # kube only?
+      accessModes: # kube only
+      - ReadWriteMany
+      size: 1Gi # kube only
+
+    env: # Optional: Ability to map key/value pairs into the container as environment variables, shared between both runtime and tasks
+    - key: value 
+    - key2: value2 
+
+  tasks:
+    - name: maven-build
+      buildImage: docker.io/maven:3.6
+      command: /scripts2/build.sh # could also just be a normal command ala `mvn clean package`
+      workingDirectory: /codewind-workspace-mount-point # optional, where in the container to run the command
+
+      logs: # Ability to reference arbitrary log file types that aren't included in container stderr/stdout
+      - type: maven.build
+        path: /logs/(etc)
+    
+      volumeMappings: #  Optional: ability to map paths in the container to persistent volume paths
+      - volumeName: idp-data-volume
+        containerPath: /idp-data-voume
+      # Map a directory for the build job to be able to copy the .war file
+      - volumeName: maven-cache-volume
+        containerPath: /home/default/.m2       
+      # This allows the .m2 cache to be persisted across job, so it doesn't need to be redownloaded every time.
+
+      repoMappings: # Optional: Automatically upload files/directories from the IDP repo to a container on/before startup
+      - srcPath: "/scripts" # path in remote git repo, where folder containing "idp.yaml" is /
+        destPath: "/home/user" # path inside container to upload the directory
+        setExecuteBit: true # Set execute bit on all files in the directory (required for windows local, git repos without execute, http serving)
+      - srcPath: "/scripts2/build.sh"
+        destPath: "/home/user/build-scripts"
+        setExecuteBit: true # Set execute bit on a single file
+
+      sourceMappings: # Optional: Ability to map files in the local project directory (eg the user's current working dir)` into the container
+      - srcPath: "/src" # copy from $CURRENT_DIR/src
+        destPath: "/home/user/src" # path inside container to copy the folder
+        setExecuteBit: true # Set execute bit on all files in the directory
+      # This is used to know where the source files should be copied into the container, might be useful for other scenarios like customization
+              
+      env: # Optional key/value env var pairs, as above
+
+      runAsUser: 185 # Optional, same as above
+      
+      kubernetes: # Optional, as above
+        livenessProbe: 
+        readinessProbe:
+        memoryLimit: 600Mi # As above
+      
+    - name: server-start
+      buildImage: docker.io/maven:3.6
+      command: /opt/ibm/wlp/bin/server start $SERVER 
+
+  scenarios:
+    - name: full-build
+      tasks: ["maven-build", "server-start"]
+    - name: incremental-build
+      tasks: ["incremental-maven-build", "server-start"]
